Travis: rename Java folder only for JREs
Soon, we will start shipping OpenJDK JDKs rather than only JREs.
As of this writing, we now do this for macOS. The script should only
attempt to rename the Java folder when the string "jre" appears in the
folder name. E.g., we want:

  jre1.8.0_172.jre -> jdk1.8.0_172.jre

But we also want to support:

  adoptopenjdk-8.jdk

Without attempting a rename.