Run the ImageJ updater twice
This is a workaround for the Updater pulling in an ancient Info.plist
the first time, which references ImageJ-tiger instead of ImageJ-macosx.
I didn't track down why that is happening, but running the update a
second time avoids the issue. And later, we will change how the
installation is generated anyway, so it's hopefully not very important.