Temporarily disable caching
The Travis GUI and CLI are reporting no caches present, but there should
be a cache. However, a cache is clearly being used because a populated
`Fiji.app` is present.

As Travis will not let us delete a cache it is unaware of, hopefully
building with a disabled cache will refresh things.