bootstrap-fiji: just use the latest fiji-nojre
It comes with launchers and native libraries of all supported platforms
already present, so that the Updater will subsequently update them
across all platforms. Whereas with the populate-app approach, those
native libraries are not bootstrapped, and get pulled down only for the
current platform when update-fiji.sh is called, which causes problems
later when the jars subfolders for other platforms are missing.

Unfortunately, this change removes the configurability of the Fiji.app
destination for the bootstrap-fiji.sh script, because unzipping the zip
file always creates a directory called Fiji.app, and I don't want to
bend over backwards to rename/move things around afterward.