update-fiji.sh: keep going if exit code 18
To avoid this issue on CI:

  [LINUX:main] Running XInitThreads
  Error: Could not find X11 library, not running XInitThreads.
  [JAUNCH:directives] INIT_THREADS directive failed with code 18

Newer fiji.toml files skip the `INIT_THREADS` when running headless,
but we have a chicken-and-egg problem here: we would receive a
working fiji.toml from the update, but the update exits non-zero...

So for now, let's just keep going on exit code 18 as well.
This hack can be removed later, after Fiji-Latest and Fiji-Stable
both have new-enough versions of fiji.toml.